# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cuidam.os API Documentation",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Auth Login",
                "description": "Login an existing account",
                "operationId": "auth_login",
                "requestBody": {
                    "description": "The request body receives email and password",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "email@email.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "YouAmazingPasswordWithMinimunLenght8SpecialCharacterCapitalLetter"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Auth Data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJnRFdxSnM3M0tQWlhLYjhqUU9tZTFadkNIemsyIiwiaXNzIjoiaHR0cDovL2N1aWRhbS5vcyIsImlhdCI6MTY2ODg2NjUwNiwiZXhwIjoxNjY4OTUyOTA2LCJuYmYiOjE2Njg4NjY1MDZ9.wY8VmD--_wln_UN6bW3AiSYiv5F9s-P2H0NiwCklSBk"
                                                },
                                                "email_link_confirmation": {
                                                    "type": "string",
                                                    "example": "https://cuidamos-91643.firebaseapp.com/__/auth/action?mode=verifyEmail&oobCode=4Rg_o4AQMQEDIIAkJHWZLqctXNa8QGcdq39TmSMtFtUAAAGEkDH9nQ&apiKey=AIzaSyAMbxYok6O6NrdTkGb-O47TObrx1DUTjFw&lang=en"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {}
                ]
            }
        },
        "/api/auth/create": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Auth Create",
                "description": "Create a new account",
                "operationId": "auth_create",
                "requestBody": {
                    "description": "The request body receives email and password and a password confirmation",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "email@email.com"
                                    },
                                    "password": {
                                        "description": "The password must have at least 8 characters, one capital letter, one number and one special character",
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "YouAmazingPasswordWithMinimunLenght8SpecialCharacterCapitalLetter"
                                    },
                                    "password_confirmation": {
                                        "description": "The password confirmation must be the same as the password",
                                        "type": "string",
                                        "maxLength": 100,
                                        "example": "YouAmazingPasswordWithMinimunLenght8SpecialCharacterCapitalLetter"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Auth Data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJnRFdxSnM3M0tQWlhLYjhqUU9tZTFadkNIemsyIiwiaXNzIjoiaHR0cDovL2N1aWRhbS5vcyIsImlhdCI6MTY2ODg2NjUwNiwiZXhwIjoxNjY4OTUyOTA2LCJuYmYiOjE2Njg4NjY1MDZ9.wY8VmD--_wln_UN6bW3AiSYiv5F9s-P2H0NiwCklSBk"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {}
                ]
            }
        },
        "/api/auth/logout": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Auth Logout",
                "description": "Logout an logged account",
                "operationId": "auth_logout",
                "responses": {
                    "200": {
                        "description": "Auth Data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Successfully logged out"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "cuidamos_auth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "cuidamos_auth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}